Remove createEntityRef method, replacing it with getEntity
PojoEntityManager's createEntityRef method was duplicating
functionality from the EntityPool's getEntity method. It also meant
that there were multiple ways that EntityRefs were created, so code
had to be updated in multiple places.

This led to bugs, such as entities not being assigned to a pool when
they are created, because the code wasn't updated in all places.

This commit removes the createEntityRef method, and replaecs all
references to it with the getEntity method.